refactor: default sticky entity page layout
