+++ packages/app-next/src/App.tsx
@@ -44,7 +44,6 @@ import kubernetesPlugin from '@backstage/plugin-kubernetes/alpha';
 import { convertLegacyPlugin } from '@backstage/core-compat-api';
 import { convertLegacyPageExtension } from '@backstage/core-compat-api';
 import { convertLegacyEntityContentExtension } from '@backstage/plugin-catalog-react/alpha';
-import { customEntityContentOverviewLayoutModule } from './EntityPages';
 
 /*
 
@@ -132,7 +131,6 @@ const app = createApp({
     kubernetesPlugin,
     notFoundErrorPageModule,
     customHomePageModule,
-    customEntityContentOverviewLayoutModule,
     ...collectedLegacyPlugins,
   ],
   /* Handled through config instead */
